pyproject.toml
[tool.poetry]
name = "config"
version = "0.0.0"
description = "Шаблон для проектов на Django"
authors = ["Azreil-OFD <fantom2413@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
gunicorn = "^21.2.0"
waitress = "^3.0.0"
django-environ = "^0.11.2"
whitenoise = "^6.6.0"
django-cors-headers = "^4.3.1"
dj-static = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
project = "config:main"